requirements.txt
langchain==0.3.3
langchain-ollama==0.2.0
matplotlib==3.9.2
ollama==0.3.3
pandas==2.2.3
requests==2.32.3
seaborn==0.13.2
fpdf2==2.8.1
pytest==8.3.3
dbt-duckdb>=1.9.0
duckdb>=1.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0